+++ civulator/__init__.py
@@ -1,3 +1,3 @@
 # Civulator - RL environment for Civilization-like strategy game
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"